requirements.txt
awscrt==0.31.3
awsiotsdk==1.28.2
boto3==1.34.65
botocore==1.34.162
jmespath==1.1.0
numpy==2.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
urllib3==2.6.3
